newly implemented methods
Here I implemented the nodeat/prettyprint/size for the different
classes either by writing it in the superclass or in the class itself.
Nokids should probably be destroyed because its only used on Num. The
point of the nodeat is mostly to make it easy to get a list of the
nodes in the tree.