Mutation needs
I didn’t really change much, but I started to work on giving the
Program a mutate() method. In the comments I wrote some of the
information it would be nice to have (thats not clear, but if you read
the comments on mutate() hopefully it will become more clear). If you
could supply that information, that would be stellar. Otherwise we can
talk about this tomorrow.