//comments added, bugs fixed
The //comment addition took way longer than I thought, and Ill admit
it’s clear that the person who wrote the rest of Tokenizer is not the
same as the person who wrote the part that handles comments, but I
believe it works (see Jparse rule2). Also I forgot to update the
pointer in the linked lists for update and rules, so they weren’t
working for linked lists greater than size 1, but I fixed that. I
changed Onekid back to its original form because there were problems
otherwise.