Mutation constructors + equality added
Each mutation has two constructors: one with a parameter for the Program
using it and one without. Certain mutations require knowing what tree
the node it's mutating is from, such as Remove and Insert. Only the
constructors with the parameter should be used for these mutations.

Each mutation is given an enum representing what type of mutation it is.
The equals method checks if the enums are the same.