merging
Did a little bit more to merge our projects and I gave you a getNumRep
to implement. Please add fields for the constants when you get the
chance (maybe as a separate class?) so that I can get rid of the
bajillion static errors that exist in ActionPacked since it recognizes
none of the constants in world at the moment.