Moved onto Exprs
I started work on interpreting the other Exprs. Unfortunately, because
MemAccess is an Expr I wound up giving pretty much all parts of the
evaluator access to the critter. I was trying to keep them as separate
as possible… Looks like that’s less separate than I thought. I still
have some actions left to cover, but a lot of the expressions were easy
to evaluate.