I moved functionality around
I gave the critter the abilities to do things like move, instead of
having these actions as static classes in the AST. This way our AST
isn’t cluttered with details about a tangential subject (Ie. the
critter). I was also hoping this would facilitate adding other
creatures into the game (which I doubt we would do, but from a design
standpoint it would be nice to have the option)