Swap/replace
I wrote swap replace methods in world so that I can use them for the
critters. I just realized we have to keep track of each hex’s position
in two places, so I decided to consolidate this in the methods swap and
replace. If there’s an error in them, let me know.