Death, testing, isingrid
I made it so that there’s a check to see if the critter dies before it
performs a specified action, but after it has lost the energy to
perform that action as per the spec. I wrote a little bit more in the
test class. I incorporated isingrid into my methods  (Mostly just my
method for checking if a space is clear and getnumrep) to make sure
that places outside of the grid are handled (i.e. treated like rocks)