Shortserve
I realized I would need to make it so I could add food to a hex, so I
decided to make use of your addfood method. I altered the return for
getNumRep of food so it returns 0 when no food is on the hex (rather
than -1 which would imply the existence of a rock).