Couple changes and an error
1. I fixed the tokenizer so it can handle // comments.
2. The critter rule files cannot have an extra line at the end. If
there is one, the rule set will not parse. Maybe if I have more time,
I’ll try to make it so that it can handle the extra line, but for now,
just delete them
3. For whatever reason, parsecritterRules needs critterReader as an
argument, not fr.

Existing problems. When a critter is loaded, the instance values for
its row and column do not match up with its given values. Not really
sure what thats about. I can’t really tell where the problem is