World gui
I have started work on the world gui. It zooms and you can move it
around, but the spacing seems like its a little off, in that when you
zoom in, im not sure it zooms directly into the center, though I’m not
sure why. Ill start adding the other graphics soon.