Pesky bugs, functional zoom
I caught some VERY pesky bugs the most elusive of which stemmed from an
unfortunate/unintentional use of dynamic dispatch. Things now zoom.
What’s left to do is to add the other types of inhabitants (critter,
food) which hopefully won’t be too bad because now that a single object
can zoom and be moved, all of the objects should be able to. The step
after adding these things is making sure the code can step through the
world model correctly, and then making sure the user can load critters
by clicking a hex… still a fair bit of work to do.