Old bugs resurface
I’ve found a couple errors that would show up in previous assignments.
I forgot a break in the critter attack sequence, so critters kept going
backwards after they attacked. A bigger deal was that new critters were
added directly to the critter list which modifies an iterator in
progress which was bad. Also, MathOp didn’t quite copy correctly
because the Op part didn’t translate in the copying. At this point
though, it looks like the world runs pretty smoothly. Just have to take
care of the observer pattern, make it so critters can be placed, and
make it so that we can get critter information.